First pass at caching grower solution
The idea is to cache both sampler and grower node solutions, to get the
grower to stick to a moving surface. This can't be done by simply caching
the geometry because it still needs to adapt to the changing sample
placement, so we do cache the barycentric coordinates in the triangles
instead (I had to disable the area weighting because that would no
longer work with stretching triangles).
On the grower node, the solution is a lot trickier: because everything is
based on spatial queries, we can't simply cache the graph. Moreover,
there's an arbitrary number of nodes interpolating between any two
attractor points. The solution is to cache the growth sequence (active
attractors, closest neighbours) and regenerate the graph every time.

The current implementation is horribly inefficient, but serves to validate
the concept.